requirements.txt
# Core Dependencies
fastapi
uvicorn[standard]
openai
sentence-transformers
faiss-cpu; sys_platform != "win32"  # On Windows, install via conda: `conda install -c conda-forge faiss-cpu`
pytest
python-dotenv
torchmetrics
transformers
huggingface-hub
requests
urllib3
certifi
tqdm
PyYAML
numpy
scipy
# torch

# Explicitly added transitive dependencies to avoid ModuleNotFound errors
scikit-learn
tokenizers
safetensors
regex
packaging
filelock
distro
httpx
idna
sniffio
anyio
typing-extensions
typing-inspection
annotated-types